+++ mne_icalabel/features/__init__.py
@@ -1,0 +1,3 @@
+"""Features for the ICLabel"""
+
+from .topomap import get_topomap_array, get_topomaps  # noqa: F401