+++ mne_icalabel/utils/tests/test_checks.py
@@ -4,7 +4,7 @@ from mne import create_info
 from mne.io import RawArray
 from mne.preprocessing import ICA
 
-from mne_icalabel.utils import _validate_inst_and_ica
+from mne_icalabel.utils._checks import _validate_inst_and_ica
 
 
 def test_validate_inst_and_ica():